requirements.txt
Flask==2.2.2
openai==0.27.6
youtube_transcript_api==0.6.0
gunicorn==20.1.0
python-dotenv==1.0.0